+++ src/main/java/marcomanfrin/atixbackend/DTO/works/WorkRequest.java
@@ -41,9 +41,9 @@ public record WorkRequest(
 
         String nasSubDirectory,
 
-        Integer expectedOfficeHours,
+        Double expectedOfficeHours,
 
-        Integer expectedPlantHours,
+        Double expectedPlantHours,
 
         UUID ticketId
 ) {
